requirements.txt
aiogram @ https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
cryptography==39.0.1
python-dotenv==1.0.0
Requests==2.31.0
rich==13.5.2
web3==6.7.0